tests/requirements.txt
selenium==4.15.0
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0
Pillow==10.1.0
opencv-python==4.8.1.78
webdriver-manager==4.0.1
requests==2.31.0
python-dotenv==1.0.0
allure-pytest==2.13.2
pytest-rerunfailures==12.0
beautifulsoup4==4.12.2
scikit-image==0.22.0
numpy==1.24.3
